safari-600.7-branch/LayoutTests/java/lc3/ArrayMethods/byte-002-expected.txt

LiveConnect 3.0 java array object inheritance JavaScript Array methods

PASS b.valueOf() is b
PASS successfullyParsed is true

TEST COMPLETE
 
